Added CheckMenuChoice class and added exeption handling for false menu choise and also made numbers to be double type.

+++ src/main/java/io/github/patrikalm/CalculationModule.java
@@ -2,23 +2,23 @@ package io.github.patrikalm;
 
 public class CalculationModule {
 
-    static int addition(int x, int y) {
+    static double addition(double x, double y) {
 
         return x + y;
 
     }
 
-    static int subtraction(int x, int y) {
+    static double subtraction(double x, double y) {
 
         return x - y;
     }
 
-    static int multiplication(int x, int y){
+    static double multiplication(double x, double y){
 
         return x * y;
     }
 
-    static int division(int x, int y) {
+    static double division( double x, double y) {
 
         return x / y;
     }
